[fuzz-server] make cycle count a 64bit value
This makes correct alignment a lot easier.
An optimization could be to combine the 16bit
cycle counts for the next 4 tests in a single 64bit field.

However, we do not have time to implement this
right now.

+++ verilator/fpga_queue.cpp
@@ -141,7 +141,7 @@ void FPGAQueueFuzzer::init(size_t coverage_size) {
 }
 bool FPGAQueueFuzzer::done() {
 	if(tests_left > 0) {
-		inputs_left = change_endianess(read_from_test<uint16_t>());
+		inputs_left = change_endianess(read_from_test<uint64_t>());
 		tests_left -= 1;
 		return false;
 	} else {
@@ -149,7 +149,7 @@ bool FPGAQueueFuzzer::done() {
 		const bool done = !acquire_buffer();
 		if(!done) {
 			parse_header();
-			inputs_left = change_endianess(read_from_test<uint16_t>());
+			inputs_left = change_endianess(read_from_test<uint64_t>());
 			tests_left -= 1;
 		} else {
 			unmap_shms();
